Opened project in nimbus, made an exercise on refs. and pointers.
Had to readd collision to the doors (seems that it was not being tracked
by the version control), and the project files were recompiled with a newer
version of the engine (4.22, I believe).

There was also an exercise on pointers and references from the lectures,
which stand as a project in the folder RefsAndPointersExample.